Cambios en tipo de Cobro

+++ src/main/java/py/com/fuentepy/appfinanzasBackend/resource/ahorro/AhorroModel.java
@@ -101,7 +101,7 @@ public class AhorroModel implements Serializable {
 
     @JsonProperty("tipo_cobro_id")
     @SerializedName("tipo_cobro_id")
-    private Integer tipoCobroId;
+    private Long tipoCobroId;
 
     @JsonProperty("entidad_financiera_nombre2")
     @SerializedName("entidad_financiera_nombre2")
