+++ src/main/java/py/com/fuentepy/appfinanzasBackend/resource/ahorro/AhorroRequestNew.java
@@ -93,5 +93,5 @@ public class AhorroRequestNew implements Serializable {
 
     @JsonProperty("tipo_cobro_id")
     @SerializedName("tipo_cobro_id")
-    private Integer tipoCobroId;
+    private Long tipoCobroId;
 }
